+++ server/app.js
@@ -1,30 +1,30 @@
 var express = require('express');
-var mysql = require('mysql');
-var connection = mysql.createConnection({
-   host     : 'database-dublease.cgiuptojslql.us-west-2.rds.amazonaws.com',
-   port     : '3306',
-   user     : 'admin',
-   password : 'Ryan6666!!',
-   database : 'dublease'
- });
 var app = express();
+
+var db = require('./data/database');
+var account = require('./user/account');
 
 app.get('/', function (req, res) {
    res.send('Hello World');
-})
+});
+
+// Sign-in Module
+app.post('/signup', async (req, res) => {
+   code, msg = account.signup(req, res);
+   res.status(code);
+   res.send(msg);
+});
+app.post('/signin', async (req, res) => {
+   code, msg = account.verify_signin(req, res);
+   res.status(code);
+   res.send(msg);
+});
 
 var server = app.listen(8000, function () {
-   var host = server.address().address
-   var port = server.address().port
+   var host = server.address().address;
+   var port = server.address().port;
    
-   console.log("Example app listening at http://%s:%s", host, port)
-})
+   console.log("Example app listening at http://%s:%s", host, port);
 
-connection.connect(function(err){
-   if(!err) {
-       console.log("Database is connected ... ");    
-   } else {
-       console.log("Error connecting database ... ");   
-       console.log(err); 
-   }
+   db.database_init();
 });
